Move AST and CFG commands to separate file
